backend/requirements.txt
# FastAPI Backend
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
pydantic>=2.0.0

# Google Cloud Secret Manager
google-cloud-secret-manager>=2.16.0

# LiveKit
livekit>=0.11.0
livekit-api>=0.6.0
livekit-agents>=0.8.0
livekit-plugins-openai>=0.8.0
livekit-plugins-google>=0.8.0
livekit-plugins-silero>=0.6.0

# OpenAI for profile extraction
openai>=1.0.0

# Google GenAI for resume parsing
google-genai>=1.0.0

# LangGraph for agentic interview prep pipeline
langgraph>=0.2.0

# Langfuse for observability (self-hosted)
langfuse>=3.0.0

# WebSocket support
websockets>=12.0
